Rename bringe to qtum_bridge
