+++ qtum_bridge/urls.py
@@ -1,6 +1,6 @@
 import tornado.web
 import settings
-from pdms.bridge import views 
+from qtum_bridge import views 
 
 bridge_router = tornado.web.Application([
         	(settings.ENDPOINTS["bridge"], views.MainHandler),
