+++ qtum_bridge/views.py
